+++ viewmodel-core/src/commonMain/kotlin/viewmodel/IntentBus.kt
@@ -1,12 +1,9 @@
-package tz.co.asoft
+package viewmodel
 
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.SupervisorJob
-import kotlinx.coroutines.delay
 import kotlinx.coroutines.flow.MutableSharedFlow
-import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.collect
-import kotlinx.coroutines.flow.filterNotNull
 import kotlinx.coroutines.launch
 
 open class IntentBus<I> {