+++ viewmodel-core/src/commonTest/kotlin/logging/ViewModelLogging.kt
@@ -2,8 +2,7 @@
 
 import TodoViewModel.Intent
 import kotlinx.coroutines.delay
-import tz.co.asoft.ConsoleAppender
-import tz.co.asoft.Logging
+import logging.*
 import tz.co.asoft.asyncTest
 import kotlin.test.Test
 
@@ -14,7 +13,7 @@ class ViewModelLogging {
     }
 
     @Test
-    fun should_print_logging_output() = asyncTest{
+    fun should_print_logging_output() = asyncTest {
         val vm = TodoViewModel()
         delay(50)
         vm.post(Intent.ReInit)