+++ viewmodel-test/viewmodel-test-expect/src/commonTest/kotlin/ApiTest.kt
@@ -4,11 +4,9 @@ import kotlinx.coroutines.coroutineScope
 import kotlinx.coroutines.launch
 import tz.co.asoft.asyncTest
 import tz.co.asoft.expect
-import tz.co.asoft.test
+import viewmodel.test
 import tz.co.asoft.toBeIn
 import kotlin.test.Test
-import kotlin.test.assertEquals
-import kotlin.test.expect
 import kotlin.time.ExperimentalTime
 import kotlin.time.measureTime