+++ CHANGELOG.md
@@ -1,12 +1,19 @@
+# 0.0.70 : 2021.02.04
+
+- changed package name from `tz.co.asoft` to `viewmodel`
+
 # 0.0.62 : 2021.01.08
-- Removed `@JsExport` from `VComponent` 
+
+- Removed `@JsExport` from `VComponent`
 
 # 0.0.61 : 2021.01.08
+
 - Updated to buidlers version 1.3.0
 - Updated and made compatible to reakt 0.0.90
 - Provided android support from jvm artifacts in `viewmodel-test-core` & `viewmodel-test-expect`
 
 # 0.0.60 : 2020.12.27
+
 - Added separated test artifacts
 - Added viewmodel-test-core artifact
 - Added viewmodel-test-expect artifact
@@ -14,36 +21,49 @@
 - made execute to be an extension of CoroutineScope
 
 # 0.0.5 : 2020.12.25
+
 - Updated to reakt version 0.0.8
 - Added viewmodel-test artifact
 - Updated to builders version 1.2.0
 
 # 0.0.4 : 2020.12.22
-- Changed signature of `post`  
+
+- Changed signature of `post`
 
 # 0.0.3 : 2020.12.19
+
 ## Updates
+
 - Updated to kotlin 1.4.21
 
 ## Bug Fixes
+
 - Fixed (Provide proper logging info on JS check on JVM as well)[https://github.com/aSoft-Ltd/viewmodel/issues/4]
 
 # 0.0.2 : 2020.12.16
+
 ## New Features
+
 - Added `VComponent` in viewmodel-react
 
 # 0.0.1
+
 ## Availability
+
 - Published to maven central
 
 ## Build Src
+
 - Updated to gradle version 6.7.1
 
 ## New Features
+
 - Initial release
 
 ## Documentation
+
 - Update readme
 
 ## Samples
+
 - Added samples